Create StrEnumWithXml to avoid duplicated code
